adding main.py to control standalone vs arch_search mode in slurm, updating GD_search to support multisearch and fixed deepcopy bug

+++ config.py
@@ -170,7 +170,7 @@ class SystemHierarchyConfig:
     self.num_nodes_per_wafer  = config_dict['num_nodes_per_wafer'] 
     #This is redundant but makes my life easier.
     self.tot_nodes            = config_dict['tot_nodes']
-    self.num_wafers           = math.ceil(self.tot_nodes / self.num_nodes_per_wafer)
+    self.num_wafers           = int(math.ceil(self.tot_nodes / self.num_nodes_per_wafer))
     self.device_placement     = ParallelMap(config_dict['device_placement'], 
                                             self.num_wafers, 
                                             self.num_nodes_per_wafer)
